Rename template packages+variables to folder. Remove all references to demo PingService
