+++ config/istio-service.yaml
@@ -1,7 +1,7 @@
 apiVersion: networking.istio.io/v1alpha3
 kind: VirtualService
 metadata:
-  name: kernel-service-poc-vservice
+  name: terra-folder-manager-vservice
 # Basic Istio virtual service to illustrate kustomize base+overlay
 # The below spec will be overwritten by the environment overlay to 
 #   namespace traffic to the right environment based on a prefix.
@@ -13,9 +13,9 @@ spec:
   http:
   - match:
     - uri:
-        prefix: /kernel-service-poc
+        prefix: /terra-folder-manager
     route:
     - destination:
         port:
           number: 8080
-        host: kernel-service-poc-service.default.svc.cluster.local
+        host: terra-folder-manager-service.default.svc.cluster.local
