+++ config/kustomization.yaml
@@ -1,7 +1,7 @@
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 commonLabels:
-  app: kernel-service-poc
+  app: terra-folder-manager
 resources:
 - service-account.yaml
 - psp.yaml
