+++ config/role.yaml
@@ -1,13 +1,13 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: Role
 metadata:
-  name: kernel-service-poc-service-role
+  name: terra-folder-manager-service-role
 rules:
 - apiGroups: ['policy']
   resources: ['podsecuritypolicies']
   verbs:     ['use']
   resourceNames:
-  - kernel-service-poc-service-psp
+  - terra-folder-manager-service-psp
 - apiGroups: ['']
   resources: ['secrets']
   verbs: ['get', 'create']
