+++ config/service-account.yaml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: kernel-service-poc-service-sa
+  name: terra-folder-manager-service-sa
