+++ config/service.yaml
@@ -1,10 +1,10 @@
 apiVersion: "v1"
 kind: "Service"
 metadata:
-  name: "kernel-service-poc-service"
+  name: "terra-folder-manager-service"
 spec:
   ports:
   - name: "http"
     port: 8080
   selector:
-    app: "kernel-service-poc"
+    app: "terra-folder-manager"
