+++ src/main/java/bio/terra/folder/app/Main.java
@@ -1,4 +1,4 @@
-package bio.terra.TEMPLATE.app;
+package bio.terra.folder.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class})
-@ComponentScan(basePackages = "bio.terra.TEMPLATE")
+@ComponentScan(basePackages = "bio.terra.folder")
 public class Main {
   public static void main(String[] args) {
     SpringApplication.run(Main.class, args);
