+++ src/main/java/bio/terra/folder/app/configuration/ApiResourceConfig.java
@@ -1,4 +1,4 @@
-package bio.terra.TEMPLATE.app.configuration;
+package bio.terra.folder.app.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
