+++ src/main/java/bio/terra/folder/common/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package bio.terra.TEMPLATE.common.exception;
+package bio.terra.folder.common.exception;
 
 // This base class has data that corresponds to the ErrorReport model generated from
 // the OpenAPI yaml. The global exception handler auto-magically converts exceptions
@@ -7,26 +7,26 @@ package bio.terra.TEMPLATE.common.exception;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 
-public abstract class NotFoundException extends ErrorReportException {
-  private static final HttpStatus thisStatus = HttpStatus.NOT_FOUND;
+public abstract class ConflictException extends ErrorReportException {
+  private static final HttpStatus thisStatus = HttpStatus.CONFLICT;
 
-  public NotFoundException(String message) {
+  public ConflictException(String message) {
     super(message, null, thisStatus);
   }
 
-  public NotFoundException(String message, Throwable cause) {
+  public ConflictException(String message, Throwable cause) {
     super(message, cause, null, thisStatus);
   }
 
-  public NotFoundException(Throwable cause) {
+  public ConflictException(Throwable cause) {
     super(null, cause, null, thisStatus);
   }
 
-  public NotFoundException(String message, List<String> causes) {
+  public ConflictException(String message, List<String> causes) {
     super(message, causes, thisStatus);
   }
 
-  public NotFoundException(String message, Throwable cause, List<String> causes) {
+  public ConflictException(String message, Throwable cause, List<String> causes) {
     super(message, cause, causes, thisStatus);
   }
 }
