+++ src/main/java/bio/terra/folder/common/exception/NotImplementedException.java
@@ -1,4 +1,4 @@
-package bio.terra.TEMPLATE.common.exception;
+package bio.terra.folder.common.exception;
 
 // This base class has data that corresponds to the ErrorReport model generated from
 // the OpenAPI yaml. The global exception handler auto-magically converts exceptions
@@ -7,26 +7,26 @@ package bio.terra.TEMPLATE.common.exception;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 
-public abstract class UnauthorizedException extends ErrorReportException {
-  private static final HttpStatus thisStatus = HttpStatus.UNAUTHORIZED;
+public abstract class NotImplementedException extends ErrorReportException {
+  private static final HttpStatus thisStatus = HttpStatus.NOT_IMPLEMENTED;
 
-  public UnauthorizedException(String message) {
+  public NotImplementedException(String message) {
     super(message, null, thisStatus);
   }
 
-  public UnauthorizedException(String message, Throwable cause) {
+  public NotImplementedException(String message, Throwable cause) {
     super(message, cause, null, thisStatus);
   }
 
-  public UnauthorizedException(Throwable cause) {
+  public NotImplementedException(Throwable cause) {
     super(null, cause, null, thisStatus);
   }
 
-  public UnauthorizedException(String message, List<String> causes) {
+  public NotImplementedException(String message, List<String> causes) {
     super(message, causes, thisStatus);
   }
 
-  public UnauthorizedException(String message, Throwable cause, List<String> causes) {
+  public NotImplementedException(String message, Throwable cause, List<String> causes) {
     super(message, cause, causes, thisStatus);
   }
 }
