+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
 server.port=8080
-template.dbUri=jdbc:postgresql://127.0.0.1:5432/${DATABASE_NAME}
-template.dbUsername=${DATABASE_USER}
-template.dbPassword=${DATABASE_USER_PASSWORD}
-template.dbInitializeOnStart=true
-template.dbInitialize=false
-template.dbUpgradeOnStart=true
+folder.dbUri=jdbc:postgresql://127.0.0.1:5432/${DATABASE_NAME}
+folder.dbUsername=${DATABASE_USER}
+folder.dbPassword=${DATABASE_USER_PASSWORD}
+folder.dbInitializeOnStart=true
+folder.dbUpgradeOnStart=true
