+++ src/test/resources/application.properties
@@ -1,6 +1,6 @@
 server.port=8080
-template.dbUri=jdbc:postgresql://127.0.0.1:5432/foldertest
-template.dbUsername=folderuser
-template.dbPassword=folderpassword
-template.dbInitialize=true
-template.dbUpgradeOnStart=true
+folder.dbUri=jdbc:postgresql://127.0.0.1:5432/foldertest
+folder.dbUsername=folderuser
+folder.dbPassword=folderpassword
+folder.dbInitializeOnStart=true
+folder.dbUpgradeOnStart=true
